chatbot/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pyttsx3
python-dotenv
flask==3.0.0
flask-cors==4.0.0
twilio